Get rid of separate views modules for collections
